+++ hq/app/services/IamRemediationService.scala
@@ -1,14 +1,13 @@
 package services
 
 import aws.AwsClients
-import aws.iam.IAMClient
 import aws.s3.S3.getS3Object
 import com.gu.janus.JanusConfig
-import config.Config.*
+import config.Config
+import config.Config.{getAnghammaradSNSTopicArn, getIamDynamoTableName, getIamUnrecognisedUserConfig}
 import db.IamRemediationDb
-import logic.IamOutdatedCredentials.*
 import logic.IamUnrecognisedUsers.*
-import model.*
+import logic.{IamOutdatedCredentials, IamUnrecognisedUsers}
 import notifications.AnghammaradNotifications
 import org.joda.time.{DateTime, DateTimeConstants}
 import play.api.inject.ApplicationLifecycle
@@ -42,61 +41,6 @@ class IamRemediationService(
 )(implicit ec: ExecutionContext)
     extends Scheduler {
 
-  /** If an AWS access key has not been rotated in a long time, then will automatically disable it.
-    *
-    * This job will first send a warning notification when it detects an outdated credential. If nothing changes it will
-    * send a final warning notification. After both these notifications have been ignored, the credential will be
-    * automatically disabled.
-    */
-  def disableOutdatedCredentials()(implicit ec: ExecutionContext): Attempt[Unit] = {
-    val now = new DateTime()
-    val result = for {
-      // lookup essential configuration
-      notificationTopicArn <- getAnghammaradSNSTopicArn(config)
-      tableName <- getIamDynamoTableName(config)
-      serviceAccountIds <- getAccountsForIamRemediationService(config)
-      allowedAwsAccountIds <- getAllowedAccountsForStage(
-        config
-      ) // this tells us which AWS accounts we are allowed to make changes to
-      // fetch IAM data from the application cache
-      rawCredsReports = cacheService.getAllCredentials
-      accountsCredReports = getCredsReportDisplayForAccount(rawCredsReports)
-      // identify users with outdated credentials for each account, from the credentials report
-      accountUsersWithOutdatedCredentials = identifyAllUsersWithOutdatedCredentials(accountsCredReports, now)
-      // DB lookup of previous SHQ activity for each user to produce a list of "candidate" vulnerabilities
-      vulnerabilitiesWithRemediationHistory <- lookupActivityHistory(
-        accountUsersWithOutdatedCredentials,
-        dynamo,
-        tableName
-      )
-      // based on activity history, decide which of these candidates have outstanding SHQ operations
-      outstandingOperations = calculateOutstandingAccessKeyOperations(vulnerabilitiesWithRemediationHistory, now)
-      // we'll only perform operations on accounts that have been configured as eligible
-      filteredOperations = partitionOperationsByAllowedAccounts(
-        outstandingOperations,
-        allowedAwsAccountIds,
-        serviceAccountIds
-      )
-      // we won't execute these operations, but can log them instead
-      _ = filteredOperations.operationsOnAccountsThatAreNotAllowed.foreach(dummyOperation)
-      // now we know what operations need to be performed, so let's run each of those
-      results <- Attempt.traverse(filteredOperations.allowedOperations)(
-        performRemediationOperation(_, now, notificationTopicArn, tableName)
-      )
-    } yield results
-    result.tap {
-      case Left(failedAttempt) =>
-        logger.error(
-          s"Failure during 'disable outdated credentials' job: ${failedAttempt.logMessage}",
-          failedAttempt.firstException.orNull // make sure the exception goes into the log, if present
-        )
-      case Right(operationIds) =>
-        logger.info(
-          s"Successfully completed 'disable outdated credentials' job, with ${operationIds.length} operations"
-        )
-    }.unit
-  }
-
   /** Removes AWS access for colleagues that have departed.
     *
     * This feature is targeted at "recovery access", where teams keep one or two IAM users that can be used to gain
@@ -107,7 +51,7 @@ class IamRemediationService(
     * with google identity tags. If we find an IAM user tagged with an identity that is not in Janus, we can assume they
     * have left and disable the IAM user.
     */
-  def disableUnrecognisedUsers()(implicit ec: ExecutionContext): Attempt[Unit] = {
+  private def disableUnrecognisedUsers()(implicit ec: ExecutionContext): Attempt[Unit] = {
     val result = for {
       config <- getIamUnrecognisedUserConfig(config)
       // fetch and parse our stored Janus config to use the canonical source of "recognised" usernames
@@ -115,7 +59,7 @@ class IamRemediationService(
       janusData = JanusConfig.load(makeFile(s3Object.mkString))
       janusUsernames = getJanusUsernames(janusData)
       // look up the credentials report from the cache service as our source of current IAM users
-      accountCredsReports = getCredsReportDisplayForAccount(cacheService.getAllCredentials)
+      accountCredsReports = IamUnrecognisedUsers.getCredsReportDisplayForAccount(cacheService.getAllCredentials)
       // determine the unrecognised users by comparing Janus usernames to the IAM users (and filter to allowed accounts)
       allowedAccountsUnrecognisedUsers = unrecognisedUsersForAllowedAccounts(
         accountCredsReports,
@@ -143,79 +87,6 @@ class IamRemediationService(
     }.unit
   }
 
-  /** Performs the specified operation, which will be one of:
-    *   - send a warning
-    *   - send a final warning
-    *   - disable an IAM credential and send a notification that this has been done
-    *   - remove an IAM password and send a notification that this has been done
-    */
-  def performRemediationOperation(
-      remediationOperation: RemediationOperation,
-      now: DateTime,
-      notificationTopicArn: String,
-      tableName: String
-  )(implicit ec: ExecutionContext): Attempt[String] = {
-    val awsAccount = remediationOperation.vulnerableCandidate.awsAccount
-    val iamUser = remediationOperation.vulnerableCandidate.iamUser
-    val problemCreationDate = remediationOperation.problemCreationDate
-    // if successful, this record will be added to the database
-    val thisRemediationActivity = IamRemediationActivity(
-      awsAccount.id,
-      iamUser.username,
-      now,
-      remediationOperation.iamRemediationActivityType,
-      remediationOperation.iamProblem,
-      remediationOperation.problemCreationDate
-    )
-
-    (remediationOperation.iamRemediationActivityType, remediationOperation.iamProblem) match {
-      // Outdated credentials
-      case (Warning, OutdatedCredential) =>
-        val notification =
-          AnghammaradNotifications.outdatedCredentialWarning(awsAccount, iamUser, problemCreationDate, now)
-        for {
-          snsId <- AnghammaradNotifications.send(notification, notificationTopicArn, snsClient)
-          _ <- dynamo.writeRemediationActivity(thisRemediationActivity, tableName)
-        } yield snsId
-
-      case (FinalWarning, OutdatedCredential) =>
-        val notification =
-          AnghammaradNotifications.outdatedCredentialFinalWarning(awsAccount, iamUser, problemCreationDate, now)
-        for {
-          snsId <- AnghammaradNotifications.send(notification, notificationTopicArn, snsClient)
-          _ <- dynamo.writeRemediationActivity(thisRemediationActivity, tableName)
-        } yield snsId
-
-      case (Remediation, OutdatedCredential) =>
-        val notification =
-          AnghammaradNotifications.outdatedCredentialRemediation(awsAccount, iamUser, problemCreationDate)
-        for {
-          // disable the correct credential
-          userCredentialInformation <- IAMClient.listUserAccessKeys(awsAccount, iamUser, iamClients)
-          credentialToDisable <- lookupCredentialId(problemCreationDate, userCredentialInformation)
-          _ <- IAMClient.disableAccessKey(
-            awsAccount,
-            credentialToDisable.username,
-            credentialToDisable.accessKeyId,
-            iamClients
-          )
-          // send a notification to say this is what we have done
-          notificationId <- AnghammaradNotifications.send(notification, notificationTopicArn, snsClient)
-          // save a record of the change
-          _ <- dynamo.writeRemediationActivity(thisRemediationActivity, tableName)
-        } yield notificationId
-    }
-  }
-
-  /** We only perform actions on accounts that are explicitly allowed, but it is helpful to log the operation that
-    * *would* have been performed, if allowed.
-    */
-  def dummyOperation(remediationOperation: RemediationOperation): Unit = {
-    val awsAccount = remediationOperation.vulnerableCandidate.awsAccount
-    logger.warn(s"Remediation operation skipped because ${awsAccount.id} is not configured for remediation")
-    logger.warn(s"Skipping remediation action: ${formatRemediationOperation(remediationOperation)}")
-  }
-
   if (environment.mode != Mode.Test) {
     // Schedule the observable on weekdays only as we may make changes in accounts that affect live systems
     // if warnings are not heeded. Initial delay of 10 minutes, so that the cache service has time to populate
@@ -230,11 +101,26 @@ class IamRemediationService(
           (now.getHourOfDay == 14 && now.getMinuteOfHour == 0)
 
         if (isWeekday && isTimeToRun) {
-          disableOutdatedCredentials()
+          disableOutdatedCredentials
           disableUnrecognisedUsers()
         }
       }
 
     lifecycle.addStopHook(iamRemediationServiceSubscription)
   }
+
+  private def disableOutdatedCredentials = for {
+    notificationTopicArn <- getAnghammaradSNSTopicArn(config)
+    tableName <- getIamDynamoTableName(config)
+    serviceAccountIds <- Config.getAccountsForIamRemediationService(config)
+    rawCredsReports = cacheService.getAllCredentials
+    // this tells us which AWS accounts we are allowed to make changes to
+    allowedAwsAccountIds <- Config.getAllowedAccountsForStage(config)
+  } yield new IamOutdatedCredentials(snsClient, iamClients, dynamo).disableOutdatedCredentials(
+    notificationTopicArn,
+    tableName,
+    serviceAccountIds,
+    rawCredsReports,
+    allowedAwsAccountIds
+  )
 }
