+++ iam-outdated-credentials/src/main/scala/logic/IamOutdatedCredentials.scala
@@ -1,21 +1,162 @@
 package logic
 
+import aws.AwsClients
+import aws.iam.IAMClient
 import com.typesafe.scalalogging.LazyLogging
 import config.CoreConfig
 import config.CoreConfig.{daysBetweenFinalNotificationAndRemediation, daysBetweenWarningAndFinalNotification}
 import db.IamRemediationDb
+import logic.IamOutdatedCredentials.*
+import logic.IamUnrecognisedUsers.*
 import model.*
-import org.joda.time.{DateTime, Days}
+import notifications.AnghammaradNotifications
+import org.joda.time.DateTime
+import software.amazon.awssdk.services.iam.IamAsyncClient
+import software.amazon.awssdk.services.sns.SnsAsyncClient
 import utils.attempt.{Attempt, FailedAttempt, Failure}
 
 import scala.concurrent.ExecutionContext
 
+class IamOutdatedCredentials(
+    snsClient: SnsAsyncClient,
+    iamClients: AwsClients[IamAsyncClient],
+    dynamo: IamRemediationDb
+) extends LazyLogging {
+
+  /** Performs the specified operation, which will be one of:
+    *   - send a warning
+    *   - send a final warning
+    *   - disable an IAM credential and send a notification that this has been done
+    *   - remove an IAM password and send a notification that this has been done
+    */
+  private def performRemediationOperation(
+      remediationOperation: RemediationOperation,
+      now: DateTime,
+      notificationTopicArn: String,
+      tableName: String
+  )(implicit ec: ExecutionContext): Attempt[String] = {
+    val awsAccount = remediationOperation.vulnerableCandidate.awsAccount
+    val iamUser = remediationOperation.vulnerableCandidate.iamUser
+    val problemCreationDate = remediationOperation.problemCreationDate
+    // if successful, this record will be added to the database
+    val thisRemediationActivity = IamRemediationActivity(
+      awsAccount.id,
+      iamUser.username,
+      now,
+      remediationOperation.iamRemediationActivityType,
+      remediationOperation.iamProblem,
+      remediationOperation.problemCreationDate
+    )
+
+    (remediationOperation.iamRemediationActivityType, remediationOperation.iamProblem) match {
+      // Outdated credentials
+      case (Warning, OutdatedCredential) =>
+        val notification =
+          AnghammaradNotifications.outdatedCredentialWarning(awsAccount, iamUser, problemCreationDate, now)
+        for {
+          snsId <- AnghammaradNotifications.send(notification, notificationTopicArn, snsClient)
+          _ <- dynamo.writeRemediationActivity(thisRemediationActivity, tableName)
+        } yield snsId
+
+      case (FinalWarning, OutdatedCredential) =>
+        val notification =
+          AnghammaradNotifications.outdatedCredentialFinalWarning(awsAccount, iamUser, problemCreationDate, now)
+        for {
+          snsId <- AnghammaradNotifications.send(notification, notificationTopicArn, snsClient)
+          _ <- dynamo.writeRemediationActivity(thisRemediationActivity, tableName)
+        } yield snsId
+
+      case (Remediation, OutdatedCredential) =>
+        val notification =
+          AnghammaradNotifications.outdatedCredentialRemediation(awsAccount, iamUser, problemCreationDate)
+        for {
+          // disable the correct credential
+          userCredentialInformation <- IAMClient.listUserAccessKeys(awsAccount, iamUser, iamClients)
+          credentialToDisable <- lookupCredentialId(problemCreationDate, userCredentialInformation)
+          _ <- IAMClient.disableAccessKey(
+            awsAccount,
+            credentialToDisable.username,
+            credentialToDisable.accessKeyId,
+            iamClients
+          )
+          // send a notification to say this is what we have done
+          notificationId <- AnghammaradNotifications.send(notification, notificationTopicArn, snsClient)
+          // save a record of the change
+          _ <- dynamo.writeRemediationActivity(thisRemediationActivity, tableName)
+        } yield notificationId
+    }
+  }
+
+  /** We only perform actions on accounts that are explicitly allowed, but it is helpful to log the operation that
+    * *would* have been performed, if allowed.
+    */
+  private def dummyOperation(remediationOperation: RemediationOperation): Unit = {
+    val awsAccount = remediationOperation.vulnerableCandidate.awsAccount
+    logger.warn(s"Remediation operation skipped because ${awsAccount.id} is not configured for remediation")
+    logger.warn(s"Skipping remediation action: ${formatRemediationOperation(remediationOperation)}")
+  }
+
+  /** If an AWS access key has not been rotated in a long time, then will automatically disable it.
+    *
+    * This job will first send a warning notification when it detects an outdated credential. If nothing changes it will
+    * send a final warning notification. After both these notifications have been ignored, the credential will be
+    * automatically disabled.
+    */
+  def disableOutdatedCredentials(
+      notificationTopicArn: String,
+      tableName: String,
+      serviceAccountIds: List[String],
+      rawCredsReports: Map[AwsAccount, Either[FailedAttempt, CredentialReportDisplay]],
+      allowedAwsAccountIds: List[String]
+  )(implicit ec: ExecutionContext): Attempt[Unit] = {
+    val now = new DateTime()
+    val accountsCredReports = getCredsReportDisplayForAccount(rawCredsReports)
+    val accountUsersWithOutdatedCredentials = identifyAllUsersWithOutdatedCredentials(accountsCredReports, now)
+
+    val result = for {
+      // fetch IAM data from the application cache
+      // identify users with outdated credentials for each account, from the credentials report
+      // DB lookup of previous SHQ activity for each user to produce a list of "candidate" vulnerabilities
+      vulnerabilitiesWithRemediationHistory <- lookupActivityHistory(
+        accountUsersWithOutdatedCredentials,
+        dynamo,
+        tableName
+      )
+      // based on activity history, decide which of these candidates have outstanding SHQ operations
+      outstandingOperations = calculateOutstandingAccessKeyOperations(vulnerabilitiesWithRemediationHistory, now)
+      // we'll only perform operations on accounts that have been configured as eligible
+      filteredOperations = partitionOperationsByAllowedAccounts(
+        outstandingOperations,
+        allowedAwsAccountIds,
+        serviceAccountIds
+      )
+      // we won't execute these operations, but can log them instead
+      _ = filteredOperations.operationsOnAccountsThatAreNotAllowed.foreach(dummyOperation)
+      // now we know what operations need to be performed, so let's run each of those
+      results <- Attempt.traverse(filteredOperations.allowedOperations)(
+        performRemediationOperation(_, now, notificationTopicArn, tableName)
+      )
+    } yield results
+    result.tap {
+      case Left(failedAttempt) =>
+        logger.error(
+          s"Failure during 'disable outdated credentials' job: ${failedAttempt.logMessage}",
+          failedAttempt.firstException.orNull // make sure the exception goes into the log, if present
+        )
+      case Right(operationIds) =>
+        logger.info(
+          s"Successfully completed 'disable outdated credentials' job, with ${operationIds.length} operations"
+        )
+    }.unit
+  }
+
+}
 object IamOutdatedCredentials extends LazyLogging {
 
   /** Look through all credentials reports to find users with expired credentials, see below for more detail
     * (`identifyUsersWithOutdatedCredentials`).
     */
-  def identifyAllUsersWithOutdatedCredentials(
+  private def identifyAllUsersWithOutdatedCredentials(
       accountCredentialReports: List[(AwsAccount, CredentialReportDisplay)],
       now: DateTime
   ): List[(AwsAccount, List[IAMUser])] = {
@@ -63,7 +204,7 @@ object IamOutdatedCredentials extends LazyLogging {
 
   /** Given an IAMUser (in an AWS account), look up that user's activity history form the Database.
     */
-  def lookupActivityHistory(
+  private def lookupActivityHistory(
       accountIdentifiedUsers: List[(AwsAccount, List[IAMUser])],
       dynamo: IamRemediationDb,
       tableName: String
@@ -87,7 +228,7 @@ object IamOutdatedCredentials extends LazyLogging {
     * the same user could appear in the output list twice, because both of their keys may require an operation. By
     * comparing the current date with the date of the most recent activity, we know which operation to perform next.
     */
-  def calculateOutstandingAccessKeyOperations(
+  private def calculateOutstandingAccessKeyOperations(
       remediationHistories: List[IamUserRemediationHistory],
       now: DateTime
   ): List[RemediationOperation] = {
@@ -204,7 +345,7 @@ object IamOutdatedCredentials extends LazyLogging {
     * AWS accounts to run the IamRemediationService on, because not all accounts are in a ready state for this service
     * yet.
     */
-  def partitionOperationsByAllowedAccounts(
+  private def partitionOperationsByAllowedAccounts(
       operations: List[RemediationOperation],
       allowedAwsAccountIds: List[String],
       serviceAccountIds: List[String]
@@ -222,7 +363,7 @@ object IamOutdatedCredentials extends LazyLogging {
     *
     * This might fail, because it may be that no matching key exists.
     */
-  def lookupCredentialId(
+  private def lookupCredentialId(
       badKeyCreationDate: DateTime,
       userCredentials: List[CredentialMetadata]
   ): Attempt[CredentialMetadata] = {
@@ -256,7 +397,7 @@ object IamOutdatedCredentials extends LazyLogging {
     }
   }
 
-  def formatRemediationOperation(remediationOperation: RemediationOperation): String = {
+  private def formatRemediationOperation(remediationOperation: RemediationOperation): String = {
     val problem = remediationOperation.iamProblem
     val activity = remediationOperation.iamRemediationActivityType
     val username = remediationOperation.vulnerableCandidate.iamUser.username
